EventIntervals

Documentation for EventIntervals.

EventIntervals.chunk

EventIntervals.interval_levels

EventIntervals.chunk — Method

chunk(int::Interval{E, 1}, chunk_len::E [, exact::Bool = false]) where E

Break an interval, int into smaller chunks of length chunk_len. If exact is true, then any remainder of int will be dropped. Otherwise, the last chunk may not be of measure chunk_len.

EventIntervals.interval_levels — Method

interval_levels(intervals::AbstractVector{<:Interval{E, 1}}) where E

Find the "depth" of interval overlap for some intervals. Assumes sorted by onset.