API reference

Index

PointProcesses.PointProcesses

PointProcesses.AbstractPointProcess

PointProcesses.BoundedPointProcess

PointProcesses.History

PointProcesses.MarkedPoissonProcess

PointProcesses.MultivariatePoissonProcess

PointProcesses.check_residuals

PointProcesses.duration

PointProcesses.event_marks

PointProcesses.event_times

PointProcesses.ground_intensity

PointProcesses.ground_intensity_bound

PointProcesses.has_events

PointProcesses.has_events

PointProcesses.integrated_ground_intensity

PointProcesses.intensity

PointProcesses.log_intensity

PointProcesses.mark_distribution

PointProcesses.max_time

PointProcesses.min_time

PointProcesses.nb_events

PointProcesses.nb_events

PointProcesses.split_into_chunks

PointProcesses.time_change

Full docs

PointProcesses.PointProcesses — Module

A package for temporal point process modeling, simulation and inference.

PointProcesses.AbstractPointProcess — Type

AbstractPointProcess{M}

The common supertype of all temporal point processes (i.e. point processes on the real line) with mark type M.

PointProcesses.BoundedPointProcess — Type

BoundedPointProcess{M,P,T}

Store a temporal point process P with pre-defined start and end times.

PointProcesses.History — Type

History{M,T}

Linear event histories with marks of type M and locations of type T, usually times represented as real numbers.

Fields

times::Vector{T}: sorted vector of event times

marks::Vector{M}: associated vector of event marks

tmin::T: start time

tmax::T: end time

PointProcesses.MarkedPoissonProcess — Type

MarkedPoissonProcess{R,M,D}

Homogeneous temporal Poisson process with arbitrary mark distribution.

Fields

λ::R: event rate.

mark_dist::D: mark distribution with sample type M.

PointProcesses.MultivariatePoissonProcess — Type

MultivariatePoissonProcess{R}

Multivariate homogeneous temporal Poisson process.

Fields

λ::Vector{R}: event rates.

Base.append! — Method

append!(h1::History, h2::History)

Add all the events of h2 at the end of h1.

Base.push! — Method

push!(h::History, t, m)

Add event (t, m) at the end of history h.

Base.rand — Method

rand(rng, pp, tmin, tmax)

Simulate a temporal point process pp on interval [tmin, tmax) using Ogata's algorithm.

DensityInterface.logdensityof — Method

logdensityof(pp, h)

Compute the log probability density function for a temporal point process pp applied to history h:

ℓ(h) = Σₖ log λ(tₖ|hₖ) - Λ(h)

The default method uses a loop over events combined with integrated_ground_intensity, but it should be reimplemented for specific processes if faster computation is possible.

PointProcesses.check_residuals — Method

check_residuals(pp, h)

Check whether the point process pp is a good fit for history h by applying Ogata's time rescaling method. If (tₖ)ₖ is a temporal point process with intensity λ, then (Λ(tₖ))ₖ is a standard temporal Poisson process.

PointProcesses.duration — Method

duration(h::History)

Compute the difference h.tmax - h.tmin.

PointProcesses.event_marks — Method

event_marks(h::History)

Return the vector of event marks for h, sorted according to their event times.

PointProcesses.event_times — Method

event_times(h::History)

Return the sorted vector of event times for h.

PointProcesses.ground_intensity — Method

ground_intensity(pp, h, t)

Compute the ground intensity for a temporal point process pp applied to history h at time t.

The ground intensity quantifies the instantaneous risk of an event with any mark occurring at time t:

λg(t|h) = Σₘ λ(t,m|h)

PointProcesses.ground_intensity_bound — Method

ground_intensity_bound(pp, t, h)

Compute a local upper bound on the ground intensity for a temporal point process pp applied to history h at time t.

Return a tuple of the form (B, L) satisfying λg(t|h) ≤ B for all u ∈ [t, t+L).

PointProcesses.has_events — Method

has_events(h::History, tmin, tmax)

Check the presence of events in h during the interval [tmin, tmax).

PointProcesses.has_events — Method

has_events(h::History)

Check the presence of events in h.

PointProcesses.integrated_ground_intensity — Method

integrated_ground_intensity(pp, h, a, b)

Compute the integrated ground intensity (or compensator) Λ(t|h) for a temporal point process pp applied to history h on interval [a, b]:

Λ(h) = ∫ λg(t|h) dt

PointProcesses.intensity — Method

intensity(pp, m, t, h)

Compute the conditional intensity for a temporal point process pp applied to history h and event (t, m).

The conditional intensity function λ(t,m|h) quantifies the instantaneous risk of an event with mark m occurring at time t after history h.

PointProcesses.log_intensity — Method

log_intensity(pp, m, t, h)

Compute the logarithm of the conditional intensity for a temporal point process pp applied to history h and event (t, m).

PointProcesses.mark_distribution — Method

mark_distribution(pp, t, h)

Compute the distribution of marks for a temporal point process pp knowing that an event takes place at time t after history h.

PointProcesses.max_time — Method

max_time(h)

Return the end time of h (not the same as the last event time).

PointProcesses.min_time — Method

min_time(h)

Return the starting time of h (not the same as the first event time).

PointProcesses.nb_events — Method

nb_events(h::History)

Count events in h.

PointProcesses.nb_events — Method

nb_events(h::History, tmin, tmax)

Count events in h during the interval [tmin, tmax).

PointProcesses.split_into_chunks — Method

split_into_chunks(h, chunk_duration)

Split h into a vector of consecutive histories with individual duration chunk_duration.

PointProcesses.time_change — Method

time_change(h, Λ)

Apply the time rescaling t -> Λ(t) to history h.