Event history

PointProcesses.AbstractHistory — Type

AbstractHistory{L, M}

Abstract supertype for event histories with locations of type L and marks of type M.

PointProcesses.TemporalHistory — Type

TemporalHistory{M}

Linear event histories with marks of type M.

Fields

times::Vector{Float64}: vector of event times

marks::Vector{M}: vector of event marks

tmin::Float64: start time

tmax::Float64: end time

Base.push! — Method

push!(h::TemporalHistory{M}, t::Float64, m::M)

Add event (t, m) at the end of history h.

PointProcesses.duration — Method

duration(h::TemporalHistory)

Compute the difference h.tmax - h.tmin.

PointProcesses.has_events — Function

has_events(h::TemporalHistory, tmin=-Inf, tmax=Inf)

Check the presence of events in h during the interval [tmin, tmax).

PointProcesses.nb_events — Function

nb_events(h::TemporalHistory, tmin=-Inf, tmax=Inf)

Count events in h during the interval [tmin, tmax).

PointProcesses.time_change — Method

time_change(h, Λ)

Apply the time rescaling $t \mapsto \Lambda(t)$ to history h.