Point processes

PointProcesses.AbstractPointProcess — Type

AbstractPointProcess{L, M}

The common supertype of all point processes with location type L and mark type M.

Temporal point processes

PointProcesses.BoundedTemporalPointProcess — Type

BoundedTemporalPointProcess{L<:Real, M}

Store a temporal point process P with pre-defined start and end times.

PointProcesses.TemporalPointProcess — Type

TemporalPointProcess{L<:Real, M}

The common supertype of all temporal point processes (i.e. point processes on the real line) with mark type M.

PointProcesses.MultivariateTemporalPointProcess — Type

MultivariateTemporalPointProcess

The common supertype of all temporal point processes with integer marks.

PointProcesses.all_marks — Method

all_marks(pp)

Enumerate possible marks for a multivariate temporal point process.

Intensity functions

PointProcesses.ground_intensity — Method

ground_intensity(pp, h, t)

Compute the ground intensity for a temporal point process pp applied to history h at time t.

The ground intensity quantifies the instantaneous risk of an event with any mark occurring at time t[Rasmussen_2018]. It can be expressed as

\[ \lambda_g(t|h) = \sum_{m \in \mathcal{M}} \lambda(t, m|h).\]

PointProcesses.ground_intensity_bound — Method

ground_intensity_bound(pp, θ, h, t)

Compute a local upper bound on the ground intensity for a temporal point process pp applied to history h at time t[Rasmussen_2018].

Return a tuple of the form $(B, L)$ satisfying

\[ \forall u \in [t, t+L), \quad \lambda_g(t|h) \leq B\]

PointProcesses.intensity — Method

intensity(pp, h, t, m)

Compute the conditional intensity for a temporal point process pp applied to history h and event (t, m).

The conditional intensity function quantifies the instantaneous risk of an event with mark m occurring at time t[Rasmussen_2018].

PointProcesses.mark_distribution — Method

mark_distribution(pp, h, t)

Compute the distribution of marks for a temporal point process pp knowing that an event takes place at time t after history h.

Simulation

Base.rand — Method

rand(rng, pp, tmin, tmax)

Simulate a temporal point process pp on interval [tmin, tmax) using Ogata's algorithm[Ogata_1981].

Learning

Distributions.logpdf — Method

logpdf(pp, h)

Compute the log probability density function for a temporal point process pp applied to history h:

\[ \log f(h) = \sum_i \log \lambda(t_i | h_i) - \Lambda(h)\]

The default method uses a loop over events combined with integrated_ground_intensity, but it should be reimplemented for specific processes if faster computation is possible.

PointProcesses.check_residuals — Method

check_residuals(pp, h)

Check whether the point process pp is a good fit for history h by applying Ogata's time rescaling method: if $(t_i)_i$ is a temporal point process with intensity $\lambda(t)$, then $(\Lambda(t_i))_i$ is a standard temporal Poisson process.

PointProcesses.integrated_ground_intensity — Function

integrated_ground_intensity(pp, h[, t])

Compute the integrated ground intensity (or compensator) $\Lambda(t|h)$ for a temporal point process pp applied to history h:

\[ \Lambda(h) = \int \lambda_g(t|h) \mathrm{d}t.\]

The default method uses Quadrature.jl for numerical integration, but it should be reimplemented for specific processes if explicit integration is feasible.

StatsBase.fit — Method

fit(pp0, h)

Compute the optimal parameter for a temporal point process of type typeof(pp0) on history h using maximum likelihood:

\[ \hat{\theta} = \mathrm{argmax} \{f_{\theta}(h): \theta \in \Theta\}\]

The default method uses Optim.jl for numerical optimization, but it should be reimplemented for specific processes if explicit maximization is feasible.

StatsBase.params — Method

params(pp)

Retrieve point process parameters as a NamedTuple.

Rasmussen_2018Rasmussen, J. G. (2018), “Lecture Notes: Temporal Point Processes and the Conditional Intensity Function,” arXiv:1806.00221 [stat].

Ogata_1981Ogata, Y. (1981), “On Lewis’ simulation method for point processes,” IEEE Transactions on Information Theory, 27, 23–31. https://doi.org/10.1109/TIT.1981.1056305.