PointProcesses

Welcome to the documentation of PointProcesses.

PointProcesses — Module

A package for point process modeling, simulation and inference.

Quick start

To install the latest release available in the General registry, run the following code in Julia's REPL:

import Pkg
Pkg.add("PointProcesses")

To install the current development version, run this instead:

import Pkg
Pkg.add("https://github.com/gdalle/PointProcesses.jl")

If you want to get started with code examples, take a look at the Tutorial section. The objects and functions defined in this package are documented in the API reference.

Index

PointProcesses

PointProcesses.AbstractHistory

PointProcesses.AbstractPointProcess

PointProcesses.BoundedTemporalPointProcess

PointProcesses.ContinuousMarkovChain

PointProcesses.ContinuousMarkovChainStats

PointProcesses.DiscreteMarkovChain

PointProcesses.DiscreteMarkovChainStats

PointProcesses.HiddenMarkovModel

PointProcesses.InhomogeneousPoissonProcess

PointProcesses.MarkovModulatedPoissonProcess

PointProcesses.MultivariateHawkesProcess

PointProcesses.MultivariatePoissonProcess

PointProcesses.MvCategorical

PointProcesses.NaiveMultivariatePoissonProcess

PointProcesses.PoissonProcess

PointProcesses.TemporalHistory

PointProcesses.TemporalPointProcess

Base.append!

Base.push!

Base.rand

Distributions.logpdf

PointProcesses.build_transform

PointProcesses.check_residuals

PointProcesses.duration

PointProcesses.ground_intensity

PointProcesses.ground_intensity_bound

PointProcesses.has_events

PointProcesses.integrated_ground_intensity

PointProcesses.intensity

PointProcesses.mark_distribution

PointProcesses.nb_events

PointProcesses.time_change

StatsBase.fit