PointProcesses

Welcome to the documentation of PointProcesses.

PointProcesses — Module

A package for point process modeling, simulation and inference.

Quick start

To get started, take a look at the Tutorial.

Index

PointProcesses

PointProcesses.AbstractHistory

PointProcesses.AbstractPointProcess

PointProcesses.BoundedTemporalPointProcess

PointProcesses.ContinuousMarkovChain

PointProcesses.ContinuousMarkovChainStats

PointProcesses.DiscreteMarkovChain

PointProcesses.DiscreteMarkovChainStats

PointProcesses.HiddenMarkovModel

PointProcesses.InhomogeneousPoissonProcess

PointProcesses.MultivariateHawkesProcess

PointProcesses.MultivariatePoissonProcess

PointProcesses.NaiveMultivariatePoissonProcess

PointProcesses.PoissonProcess

PointProcesses.TemporalHistory

PointProcesses.TemporalPointProcess

Base.push!

Base.rand

Distributions.logpdf

PointProcesses.build_transform

PointProcesses.check_residuals

PointProcesses.duration

PointProcesses.ground_intensity

PointProcesses.ground_intensity_bound

PointProcesses.has_events

PointProcesses.integrated_ground_intensity

PointProcesses.intensity

PointProcesses.mark_distribution

PointProcesses.nb_events

PointProcesses.time_change

StatsBase.fit