API reference

Index

PointProcesses.PointProcesses

PointProcesses.BoundedTemporalPointProcess

PointProcesses.ContinuousMarkovChain

PointProcesses.ContinuousMarkovChainPrior

PointProcesses.ContinuousMarkovChainStats

PointProcesses.DiscreteMarkovChain

PointProcesses.DiscreteMarkovChainPrior

PointProcesses.DiscreteMarkovChainStats

PointProcesses.History

PointProcesses.MultivariatePoissonProcess

PointProcesses.PoissonProcess

PointProcesses.TemporalPointProcess

PointProcesses.duration

PointProcesses.ground_intensity

PointProcesses.ground_intensity_bound

PointProcesses.has_events

PointProcesses.integrated_ground_intensity

PointProcesses.intensity

PointProcesses.log_intensity

PointProcesses.mark_distribution

PointProcesses.max_time

PointProcesses.min_time

PointProcesses.nb_events

PointProcesses.simulate_ogata

PointProcesses.time_change

Full docs

PointProcesses.PointProcesses — Module

A package for point process modeling, simulation and inference.

PointProcesses.BoundedTemporalPointProcess — Type

BoundedTemporalPointProcess{M,T}

Store a temporal point process P with pre-defined start and end times.

PointProcesses.ContinuousMarkovChain — Type

ContinuousMarkovChain

Continuous-time Markov chain with finite state space.

Fields

π0: initial state distribution

Q: rate matrix.

PointProcesses.ContinuousMarkovChainPrior — Type

ContinuousMarkovChainPrior

Define a Dirichlet prior on the initial distribution and a Gamma prior on the transition rates from each state.

Fields

π0α: Dirichlet parameter

Pα: Gamma rate parameters

Pβ: Gamma shape parameters

PointProcesses.ContinuousMarkovChainStats — Type

ContinuousMarkovChainStats

Sufficient statistics for the likelihood of a ContinuousMarkovChain.

PointProcesses.DiscreteMarkovChain — Type

DiscreteMarkovChain

Discrete-time Markov chain with finite state space.

Fields

π0: initial state distribution.

P: state transition matrix.

PointProcesses.DiscreteMarkovChainPrior — Type

DiscreteMarkovChainPrior

Define a Dirichlet prior on the initial distribution and on the transitions from each state.

Fields

π0α: Dirichlet parameter for the initial distribution

Pα: Dirichlet parameters for the transition matrix

PointProcesses.DiscreteMarkovChainStats — Type

DiscreteMarkovChainStats

Sufficient statistics for the likelihood of a DiscreteMarkovChain.

PointProcesses.History — Type

History{M,T<:Real}

Linear event histories with marks of type M and locations of real type T.

Fields

times::Vector{T}: vector of event times

marks::Vector{M}: vector of event marks

tmin::T: start time

tmax::T: end time

PointProcesses.MultivariatePoissonProcess — Type

MultivariatePoissonProcess{R}

Homogeneous multivariate temporal Poisson process.

Fields

λ::Vector{R}: event rates.

PointProcesses.PoissonProcess — Type

PoissonProcess{D,M,R}

Homogeneous temporal Poisson process with arbitrary mark distribution.

Fields

λ::R: event rate.

mark_dist::D: mark distribution with sample type M.

PointProcesses.TemporalPointProcess — Type

TemporalPointProcess{M,T}

The common supertype of all temporal point processes (i.e. point processes on the real line) with mark type M and time type T.

Base.append! — Method

append!(h1::History, h2::History)

Add all the events of h2 at the end of h1.

Base.push! — Method

push!(h::History{M}, t::Float64, m::M)

Add event (t, m) at the end of history h.

DensityInterface.logdensityof — Method

logdensityof(pp, h)

Compute the log probability density function for a temporal point process pp applied to history h:

\[ \log f(h) = \sum_i \log \lambda(t_i | h_i) - \Lambda(h)\]

The default method uses a loop over events combined with integrated_ground_intensity, but it should be reimplemented for specific processes if faster computation is possible.

PointProcesses.duration — Method

duration(h::History)

Compute the difference h.tmax - h.tmin.

PointProcesses.ground_intensity — Method

ground_intensity(pp, h, t)

Compute the ground intensity for a temporal point process pp applied to history h at time t.

The ground intensity quantifies the instantaneous risk of an event with any mark occurring at time t. It can be expressed as

\[ \lambda_g(t|h) = \sum_{m \in \mathcal} \lambda(t, m|h).\]

PointProcesses.ground_intensity_bound — Method

ground_intensity_bound(pp, t, h)

Compute a local upper bound on the ground intensity for a temporal point process pp applied to history h at time t.

Return a tuple of the form $(B, L)$ satisfying

\[ \forall u \in [t, t+L), \quad \lambda_g(t|h) \leq B\]

PointProcesses.has_events — Function

has_events(h::History, tmin=-Inf, tmax=Inf)

Check the presence of events in h during the interval [tmin, tmax).

PointProcesses.integrated_ground_intensity — Function

integrated_ground_intensity(pp, h[, t])

Compute the integrated ground intensity (or compensator) $\Lambda(t|h)$ for a temporal point process pp applied to history h:

\[ \Lambda(h) = \int \lambda_g(t|h) \mathrm{d}t.\]

The default method uses Quadrature.jl for numerical integration, but it should be reimplemented for specific processes if explicit integration is feasible.

PointProcesses.intensity — Method

intensity(pp, m, t, h)

Compute the conditional intensity for a temporal point process pp applied to history h and event (t, m).

The conditional intensity function $\lambda(t, m | h)$ quantifies the instantaneous risk of an event with mark m occurring at time t.

PointProcesses.log_intensity — Method

log_intensity(pp, m, t, h)

Compute the logarithm of the conditional intensity for a temporal point process pp applied to history h and event (t, m).

PointProcesses.mark_distribution — Method

mark_distribution(pp, t, h)

Compute the distribution of marks for a temporal point process pp knowing that an event takes place at time t after history h.

PointProcesses.max_time — Method

max_time(h)

Return the end time of h (not the same as the last event time).

PointProcesses.min_time — Method

min_time(h)

Return the starting time of h (not the same as the first event time).

PointProcesses.nb_events — Function

nb_events(h::History, tmin=-Inf, tmax=Inf)

Count events in h during the interval [tmin, tmax).

PointProcesses.simulate_ogata — Method

simulate_ogata(rng, pp, tmin, tmax)

Simulate a temporal point process pp on interval [tmin, tmax) using Ogata's algorithm.

PointProcesses.time_change — Method

time_change(h, Λ)

Apply the time rescaling $t \mapsto \Lambda(t)$ to history h.