API reference

Event history

Abstract type

PointProcesses.AbstractHistory — Type

AbstractHistory{L,M}

Abstract supertype for event histories with locations of type L and marks of type M.

Temporal history

PointProcesses.TemporalHistory — Type

TemporalHistory{M}

Linear event histories with marks of type M.

Fields

times::Vector{Float64}: vector of event times

marks::Vector{M}: vector of event marks

tmin::Float64: start time

tmax::Float64: end time

Base.append! — Method

append!(h1::TemporalHistory, h2::TemporalHistory)

Add all the events of h2 at the end of h1.

Base.push! — Method

push!(h::TemporalHistory{M}, t::Float64, m::M)

Add event (t, m) at the end of history h.

PointProcesses.duration — Method

duration(h::TemporalHistory)

Compute the difference h.tmax - h.tmin.

PointProcesses.has_events — Function

has_events(h::TemporalHistory, tmin=-Inf, tmax=Inf)

Check the presence of events in h during the interval [tmin, tmax).

PointProcesses.max_time — Method

max_time(h)

Return the end time of h (not the same as the last event time).

PointProcesses.min_time — Method

min_time(h)

Return the starting time of h (not the same as the first event time).

PointProcesses.nb_events — Function

nb_events(h::TemporalHistory, tmin=-Inf, tmax=Inf)

Count events in h during the interval [tmin, tmax).

PointProcesses.time_change — Method

time_change(h, Λ)

Apply the time rescaling $t \mapsto \Lambda(t)$ to history h.

Point processes

Abstract type

PointProcesses.AbstractPointProcess — Type

AbstractPointProcess{L, M}

The common supertype of all point processes with location type L and mark type M.

PointProcesses.build_transform — Function

build_transform(pp)

Return a transformation object from TransformVariables that can turn a Vector{Float64} into a NamedTuple with fields matching those of pp.

Temporal point processes

PointProcesses.BoundedTemporalPointProcess — Type

BoundedTemporalPointProcess{L<:Real, M}

Store a temporal point process P with pre-defined start and end times.

PointProcesses.TemporalPointProcess — Type

TemporalPointProcess{L<:Real, M}

The common supertype of all temporal point processes (i.e. point processes on the real line) with mark type M.

Base.rand — Method

rand(rng, pp, tmin, tmax)

Simulate a temporal point process pp on interval [tmin, tmax) using Ogata's algorithm[Ogata_1981].

PointProcesses.check_residuals — Method

check_residuals(pp, h)

Check whether the point process pp is a good fit for history h by applying Ogata's time rescaling method: if $(t_i)_i$ is a temporal point process with intensity $\lambda(t)$, then $(\Lambda(t_i))_i$ is a standard temporal Poisson process.

PointProcesses.ground_intensity — Method

ground_intensity(pp, h, t)

Compute the ground intensity for a temporal point process pp applied to history h at time t.

The ground intensity quantifies the instantaneous risk of an event with any mark occurring at time t[Rasmussen_2018]. It can be expressed as

\[ \lambda_g(t|h) = \sum_{m \in \mathcal} \lambda(t, m|h).\]

PointProcesses.ground_intensity_bound — Method

ground_intensity_bound(pp, θ, h, t)

Compute a local upper bound on the ground intensity for a temporal point process pp applied to history h at time t[Rasmussen_2018].

Return a tuple of the form $(B, L)$ satisfying

\[ \forall u \in [t, t+L), \quad \lambda_g(t|h) \leq B\]

PointProcesses.integrated_ground_intensity — Function

integrated_ground_intensity(pp, h[, t])

Compute the integrated ground intensity (or compensator) $\Lambda(t|h)$ for a temporal point process pp applied to history h:

\[ \Lambda(h) = \int \lambda_g(t|h) \mathrm{d}t.\]

The default method uses Quadrature.jl for numerical integration, but it should be reimplemented for specific processes if explicit integration is feasible.

PointProcesses.intensity — Method

intensity(pp, h, t, m)

Compute the conditional intensity for a temporal point process pp applied to history h and event (t, m).

The conditional intensity function $\lambda(t, m | h)$ quantifies the instantaneous risk of an event with mark m occurring at time t[Rasmussen_2018].

PointProcesses.logpdf — Method

logpdf(pp, h)

Compute the log probability density function for a temporal point process pp applied to history h:

\[ \log f(h) = \sum_i \log \lambda(t_i | h_i) - \Lambda(h)\]

The default method uses a loop over events combined with integrated_ground_intensity, but it should be reimplemented for specific processes if faster computation is possible.

PointProcesses.mark_distribution — Method

mark_distribution(pp, h, t)

Compute the distribution of marks for a temporal point process pp knowing that an event takes place at time t after history h.

StatsBase.fit — Method

fit(pp0, h)

Compute the optimal parameter for a temporal point process of type typeof(pp0) on history h using maximum likelihood:

\[ \hat{\theta} = \mathrm{argmax} \{f_{\theta}(h): \theta \in \Theta\}\]

The default method uses GalacticOptim.jl for numerical optimization, but it should be reimplemented for specific processes if explicit maximization is feasible.

Built-in models

Poisson processes

PointProcesses.PoissonProcess — Type

PoissonProcess{D,M}

Homogeneous temporal Poisson process with arbitrary mark distribution.

Fields

λ::Float64: event rate.

mark_dist::D: mark distribution.

PointProcesses.NaiveMultivariatePoissonProcess — Type

NaiveMultivariatePoissonProcess{R}

Homogeneous temporal multivariate Poisson process.

This naive implementation demonstrates the use of our general TemporalPointProcess interface.

Fields

λ::Vector{R}: event rates.

Hawkes processes

Markov chains

Abstract type

Discrete time

PointProcesses.DiscreteMarkovChain — Type

DiscreteMarkovChain{T1,T2}

Discrete-time Markov chain with finite state space.

Fields

π0::T1: initial state distribution.

P::T2: state transition matrix.

PointProcesses.DiscreteMarkovChainPrior — Type

DiscreteMarkovChainPrior

Define a Dirichlet prior on the initial distribution and on the transitions from each state.

Fields

π0α: Dirichlet parameter for the initial distribution

Pα: Dirichlet parameters for the transition matrix

PointProcesses.DiscreteMarkovChainStats — Type

DiscreteMarkovChainStats

Sufficient statistics for the likelihood of a DiscreteMarkovChain.

Continuous time

PointProcesses.ContinuousMarkovChain — Type

ContinuousMarkovChain{T1,T2}

Continuous-time Markov chain with finite state space.

Fields

π0::T1: initial state distribution

Q::T2: rate matrix.

PointProcesses.ContinuousMarkovChainPrior — Type

ContinuousMarkovChainPrior

Define a Dirichlet prior on the initial distribution and a Gamma prior on the transition rates from each state.

Fields

π0α: Dirichlet parameter

Pα: Gamma rate parameters

Pβ: Gamma shape parameters

PointProcesses.ContinuousMarkovChainStats — Type

ContinuousMarkovChainStats

Sufficient statistics for the likelihood of a ContinuousMarkovChain.

Hidden Markov models

Discrete time

PointProcesses.HiddenMarkovModel — Type

HMM{Tr<:DiscreteMarkovChain,Em}

Hidden Markov Model with arbitrary transition model (must be a discrete Markov chain) and emission distributions.

Fields

transitions::Tr: state evolution process.

emissions::Vector{Em}: one emission distribution per state.

Continuous time

PointProcesses.MarkovModulatedPoissonProcess — Type

MarkovModulatedPoissonProcess{M,Tr<:ContinuousMarkovChain,Em<:TemporalPointProcess{M}}

Markov-Modulated Poisson Process with mark type M.

Fields

transitions::Tr: state evolution process.

emissions::Vector{Em}: one emission distribution per state.

Utilities

PointProcesses.plot_events — Method

plot_events(h)

Plot an event history with its marks if those are of a real subtype.

PointProcesses.plot_intensity — Method

plot_intensity(pp, h; npoints)

Plot the conditional intensity function of a temporal point process along a given history.

PointProcesses.qqplot_interevent_times — Method

qqplot_interevent_times(h)

Compare the distribution of inter-event times in a given history to the exponential distribution.

PointProcesses.randprobvec — Method

randprobvec(n)

Return a random probability distribution vector of size n.

PointProcesses.randtransmat — Method

randtransmat(n)

Return a stochastic matrix of size n with random transition probability distributions.

PointProcesses.uniformprobvec — Method

uniformprobvec(n)

Return a uniform probability distribution vector of size n.

PointProcesses.uniformtransmat — Method

uniformtransmat(n)

Return a stochastic matrix of size n with uniform transition probability distributions.

Ogata_1981Ogata, Y. (1981), “On Lewis’ simulation method for point processes,” IEEE Transactions on Information Theory, 27, 23–31. https://doi.org/10.1109/TIT.1981.1056305.

Rasmussen_2018Rasmussen, J. G. (2018), “Lecture Notes: Temporal Point Processes and the Conditional Intensity Function,” arXiv:1806.00221 [stat].